requirements-render.txt
email-validator>=2.3.0
flask-dance>=7.1.0
flask>=3.1.2
flask-migrate>=4.0.5
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
oauthlib>=3.3.1
psycopg2-binary>=2.9.10
pyjwt>=2.10.1
flask-login>=0.6.3
pandas>=2.3.2
python-dateutil>=2.9.0.post0
sqlalchemy>=2.0.43
werkzeug>=3.1.3
flask-wtf>=1.2.2
wtforms>=3.2.1
python-dotenv>=1.0.1